Finally got all of the various sub projects building on linux. There are a couple of build specific link issues with some of the embedded projects since they are much tighter on memory than the big boys are but everything at least gets to the link stage without errors.

+++ stateMachine_G4_testSuite/sm_test_calculator.c
@@ -329,7 +329,7 @@ uint8_t doCalculation(	calculatorMachine_t* sm, event_t* event)
 {
 	bool	success ;
 	char	output[50] ;
-	float	result ;
+	double	result ;
 
 	sprintf(output,
 			"%ld.%ld",
